Split collapsable section styles from core section styles. #400.
