+++ explorer/app/components/Detail/components/Section/section.styles.ts
@@ -1,6 +1,5 @@
-// TODO section styles refactor to be resolved with https://github.com/clevercanary/data-browser/issues/400
 import styled from "@emotion/styled";
-import { ButtonBase, Typography } from "@mui/material";
+import { Typography } from "@mui/material";
 import { BREAKPOINT } from "../../../../hooks/useBreakpointHelper";
 import { SectionTitle as Title } from "./components/SectionTitle/sectionTitle";
 
@@ -9,28 +8,6 @@ export const Section = styled.div`
 
   ${({ theme }) => theme.breakpoints.up(BREAKPOINT.TABLET)} {
     padding: 20px;
-  }
-`;
-
-export const CollapsableSection = styled.div`
-  background-color: ${({ theme }) => theme.palette.common.white};
-  display: flex;
-  flex-direction: column;
-  padding: 4px 0;
-
-  ${({ theme }) => theme.breakpoints.up(BREAKPOINT.TABLET)} {
-    gap: 8px;
-    padding: 20px;
-  }
-`;
-
-export const SectionSummary = styled(ButtonBase)`
-  display: flex;
-  justify-content: space-between;
-  padding: 16px;
-
-  ${({ theme }) => theme.breakpoints.up(BREAKPOINT.TABLET)} {
-    padding: 0;
   }
 `;
 
@@ -45,18 +22,6 @@ export const Content = styled(Typography)`
 
   > *:last-child {
     margin-bottom: 0;
-  }
-` as typeof Typography;
-
-export const SectionContent = styled(Typography)`
-  align-items: flex-start;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  padding: 0 16px 16px;
-
-  ${({ theme }) => theme.breakpoints.up(BREAKPOINT.TABLET)} {
-    padding: 0;
   }
 ` as typeof Typography;
